usečka

+++ src/rasterize/DottedLineRasterizer.java
@@ -11,12 +11,12 @@ public class DottedLineRasterizer extends LineRasterizer{
     }
 
     @Override
-        public void drawLine(Raster raster, double x1, double y1, double x2, double y2, int color) {
+        public void drawLine(Raster raster, int x1, int y1, int x2, int y2, int color) {
 
             int pocitadlo = 0;  //Pro mezery jsem využil počitadlo, které mně řekne,
                                 // které pixely zobrazím jako background
             if (x2 < x1) {
-                double temp = x2;
+                int temp = x2;
                 x2 = x1;
                 x1 = temp;
