+++ src/rasterize/LineRasterizerGraphics.java
@@ -15,10 +15,10 @@ public class LineRasterizerGraphics extends LineRasterizer {
     }
 
    @Override
-    public void drawLine(Raster raster, double x1, double y1, double x2, double y2, int color) {
+    public void drawLine(Raster raster, int x1, int y1, int x2, int y2, int color) {
         Graphics g = ((RasterBufferedImage)raster).getImg().getGraphics();
         g.setColor(this.color);
-        g.drawLine((int)x1, (int)y1, (int)x2, (int)y2);
+        g.drawLine(x1, y1, x2, y2);
     }
 
 }
